Modularize display.rs into submodules for maintainability
Split 609-line display.rs into focused submodules:
- time.rs: relative_time, compact_date
- terminal.rs: TTY detection, colors, width
- labels.rs: label rendering
- issues.rs: issue formatting and display
- goals.rs: goal formatting and display
- tests.rs: unit tests
- mod.rs: re-exports
